Integer arrays converted to Double arrays

+++ src/main/java/HW/App.java
@@ -5,19 +5,19 @@ package HW;
 
 
 public class App {
-    public int[] evaluate(int[] coef, int nTh, char mode)
+    public double[] evaluate(double[] coef, int nTh, char mode)
     {
         if(coef == null || (mode != 'd' && mode != 'i') || nTh < 0)
         {
             return null;
         }
 
-        int[] result = null;
+        double[] result = null;
         if(mode == 'd')
         {
             for(int i=0; i < nTh; i++)
             {
-                result = new int[coef.length-1];
+                result = new double[coef.length-1];
 
                 for(int j=1; j < coef.length; j++)
                 {
@@ -31,7 +31,7 @@ public class App {
         {
             for(int i=0; i < nTh; i++)
             {
-                result = new int[coef.length+1];
+                result = new double[coef.length+1];
 
                 for(int j=0; j < coef.length; j++)
                 {
